producer/requirements.txt
kafka-python
scikit-learn==1.2.2